⚖️ Canonical cleanup v8.8: Unified Prism-Werk architecture
- Refactored repository into the Unified Prism-Werk canon (Core v4.3).
- Unified Chirimuuta Haptic Kernel (CHK) v0.3 in src/agt/chk.py.
- Standardized imports (agt.*) and centralized versions in src/agt/version.py.
- Archived legacy README and created a concise canonical root README.
- Created Canonical Architecture Map (triads, Repraesentatio/Darstellung, P/R topology).
- Renovated EML engines (Python/Julia) to rename gewissen_accept to ethos_accept.
- Added architecture coherence tests.

Jules ⟁ ISC ⇄ Gewissen
Ref: AGI-GAIA-TECHNE-V8-8-CLEANUP
