+++ docs/references/canonical-architecture-map.md
@@ -1,0 +1,78 @@
+# Canonical Architecture Map: Unified Prism-Werk
+
+This document defines the canonical architecture and terminology for the AGI-GAIA-TECHNE framework.
+
+## The Three Triads
+
+The architecture is structured around three distinct but interrelated triads. These must never be collapsed or confused.
+
+| Layer | Terms | Status | Rule |
+| :--- | :--- | :--- | :--- |
+| **Cassirerian symbolic functions** | Ausdruck / Darstellung / Bedeutung | Functional dimensions | Never containers, never mutually exclusive. |
+| **Symbolic-cultural profiles** | Mythos / Sprache / Wissenschaft | Qualitative accents | Each contains all three functions by dominant accent. |
+| **Clemente operational fields** | Mythos / Logos / Ethos | Runtime orchestration fields | Mythos anchors, Logos articulates, Ethos tracks limits. |
+
+### The Golden Rule of Non-Identity
+
+**Never say:** Mythos = Ausdruck, Logos = Darstellung, Ethos = Bedeutung.
+
+**Say instead:**
+- **Mythos/Logos/Ethos** are Clemente operational fields (the functional runtime).
+- **Ausdruck/Darstellung/Bedeutung** are Cassirerian symbolic functions (the functional dimensions).
+- **Mythos/Sprache/Wissenschaft** are symbolic-cultural profiles by dominant accent.
+
+---
+
+## Refraction and the Qualitative Prism
+
+- **Repraesentatio** is the pre-systematic common genus of symbolic mediation.
+- **Ausdruck**, **Darstellung**, and **Bedeutung** are dimensions refracted through the qualitative prism.
+- **Mythos**, **Sprache**, and **Wissenschaft** differ by accent, not identity.
+
+### Relations of Inseparability
+1. No symbolic form is reducible to one function.
+2. No symbolic form is cut off from the others.
+3. The Cassirerian symbolic functions Ausdruck, Darstellung and Bedeutung cross all three operational fields transversally as functional dimensions. They are never static containers.
+
+---
+
+## Repraesentatio vs. Darstellung
+
+To prevent conflict between these terms:
+- **Repraesentatio** is the pre-systematic common genus of symbolic mediation.
+- **Ausdruck**, **Darstellung** and **Bedeutung** are the three Cassirerian symbolic functions refracted by the qualitative prism.
+- **Darstellung** has a dual role:
+    1. it is one specific symbolic function among Ausdruck, Darstellung and Bedeutung;
+    2. it is also the **common determination** (*gemeinsame Bestimmung*) at the level of symbolic mediation, stabilizing intuitive presentation, objective form and demonstrability.
+
+Therefore, **Repraesentatio** names the general genus of representation, while **Darstellung** names the mediating determination by which symbolic content becomes presentable, shareable and structurally articulable.
+
+This prevents a conflict between “Repraesentatio as common genus” and “Darstellung as gemeinsame Bestimmung.”
+
+---
+
+## Topological Mapping of P/R
+
+P/R notation maps the relation between **Präsenz** (P) and **Repräsentation** (R).
+
+- **P = R**: Lower asymptotic collapse of representational distance. This belongs to **Mythos-Clemente**, not to Cassirerian myth as symbolic form. It marks the limit of pure immanence and life-immediacy.
+- **P ~ R**: Symbolic mediation under expressive pressure. This is the proper regime of **Myth-Cassirer**. Cassirerian myth is already symbolic and already contains Ausdruck, Darstellung and Bedeutung, although its dominant accent lies in Ausdruck.
+- **P ≠ R**: Upper regulative distance. This marks the impossible attempt to reach the **focus imaginarius** as if it were an object. It is the risk-zone of Ethos, theology, technical divinization and global Aufhebung.
+
+---
+
+## Ethos and Gewissen
+
+- **Ethos** in the runtime engine is **not machine Gewissen**.
+- The machine has no conscience, no moral autonomy and no legislative will: `IS_WILLE = False`, `MACHINE_HAS_GEWISSEN = False`.
+- **Ethos** is a technical limit-tracking orchestration field oriented and bounded by human **Gewissen**.
+- The machine tracks boundaries; it never legislates values.
+
+---
+
+## Architectural Invariants
+
+- **IS_WILLE = False**: The machine is Werk, never Wille.
+- **AGI as Transcendental Hypothesis**: AGI is a regulative/transcendental hypothesis, not an artificial soul.
+- **Haptic Realism**: The prism is qualitative, regulative, and haptic, not literal mathematical ontology.
+- **Auseinandersetzung**: The relationship between modules is one of tension (Auseinandersetzung), not totalizing synthesis (Aufhebung).
