+++ docs/references/mythos-logos-ethos-engine.md
@@ -1,5 +1,8 @@
 # Mythos-Logos-Ethos Engine
 
+
+For the canonical terminology and triad definitions, see the [Canonical Architecture Map](docs/references/canonical-architecture-map.md).
+
 The Mythos-Logos-Ethos Engine is the functional orchestration motor of AGI-GAIA-TECHNE.
 
 It does not replace the EML kernel, the CTK or the CHK. It coordinates them architectonically.
