+++ memory/agt_memory.jsonl
@@ -4,3 +4,12 @@
 {"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk"], "created_at": 1779405081.1904273}
 {"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk"], "created_at": 1779405104.6041918}
 {"kind": "procedural", "key": "The qualitative prism is a regulative model.", "value": "Procedure used: audit → plan → execute safe tool → record.", "tags": ["procedure", "agt-controller"], "created_at": 1779405104.60897}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657750.7083015, "version": 1}
+{"kind": "procedural", "key": "The qualitative prism is a regulative model.", "value": "Procedure used: audit → step-audit → safe tool → memory.", "tags": ["procedure", "agt-controller", "werk-output"], "created_at": 1779657750.7415552, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657807.370994, "version": 1}
+{"kind": "procedural", "key": "The qualitative prism is a regulative model.", "value": "Procedure used: audit → step-audit → safe tool → memory.", "tags": ["procedure", "agt-controller", "werk-output"], "created_at": 1779657807.4017098, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657863.899699, "version": 1}
+{"kind": "procedural", "key": "The qualitative prism is a regulative model.", "value": "Procedure used: audit → step-audit → safe tool → memory.", "tags": ["procedure", "agt-controller", "werk-output"], "created_at": 1779657863.9307082, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779661913.9472113, "version": 1}
+{"kind": "procedural", "key": "The qualitative prism is a regulative model.", "value": "Procedure used: audit → step-audit → safe tool → memory.", "tags": ["procedure", "agt-controller", "werk-output"], "created_at": 1779661913.9795082, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779661914.359359, "version": 1}
