+++ memory/test_agt_memory.jsonl
@@ -1,0 +1,12 @@
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657750.5787306, "version": 1}
+{"kind": "procedural", "key": "The qualitative prism is a regulative model.", "value": "Procedure used: audit → step-audit → safe tool → memory.", "tags": ["procedure", "agt-controller", "werk-output"], "created_at": 1779657750.5832567, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657750.5844285, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657750.588547, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657807.2585738, "version": 1}
+{"kind": "procedural", "key": "The qualitative prism is a regulative model.", "value": "Procedure used: audit → step-audit → safe tool → memory.", "tags": ["procedure", "agt-controller", "werk-output"], "created_at": 1779657807.262275, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657807.2634492, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657807.2674155, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657863.787476, "version": 1}
+{"kind": "procedural", "key": "The qualitative prism is a regulative model.", "value": "Procedure used: audit → step-audit → safe tool → memory.", "tags": ["procedure", "agt-controller", "werk-output"], "created_at": 1779657863.7911623, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657863.7922106, "version": 1}
+{"kind": "normative", "key": "axioms", "value": "IS_WILLE=False; MACHINE_HAS_GEWISSEN=False; NO_GLOBAL_AUFHEBUNG=True", "tags": ["axiom", "ctk", "werk"], "created_at": 1779657863.796322, "version": 1}
