+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "agt-functional-core"
-version = "0.2.0"
+version = "0.3.0"
 requires-python = ">=3.10"
 
 [tool.pytest.ini_options]
