+++ src/agt/types.py
@@ -53,6 +53,18 @@ class ThesisStatus(str, Enum):
     CONSTITUTIVE_OVERREACH = "CONSTITUTIVE_OVERREACH"
     ABSTRACTION_COST_MISSING = "ABSTRACTION_COST_MISSING"
 
+    # CHK Rich Statuses
+    MEDIUM_DEPENDENCE_RISK = "MEDIUM_DEPENDENCE_RISK"
+    CARTESIAN_IDEALIZATION_RISK = "CARTESIAN_IDEALIZATION_RISK"
+    REFLEX_ATOMISM_RISK = "REFLEX_ATOMISM_RISK"
+    MACHINE_ORGANISM_ANALOGY_RISK = "MACHINE_ORGANISM_ANALOGY_RISK"
+    HERACLITEAN_PLASTICITY_RISK = "HERACLITEAN_PLASTICITY_RISK"
+    PREDICTION_WITHOUT_UNDERSTANDING = "PREDICTION_WITHOUT_UNDERSTANDING"
+    TECHNOCRATIC_AUTHORITY_RISK = "TECHNOCRATIC_AUTHORITY_RISK"
+    APOCALYPTIC_TECHNOLOGY_RISK = "APOCALYPTIC_TECHNOLOGY_RISK"
+    JUDGMENT_GAP = "JUDGMENT_GAP"
+    ABSTRACTION_RISK = "ABSTRACTION_RISK"
+
 
 class Pillar(str, Enum):
     MYTHOS = "Mythos"
@@ -74,6 +86,7 @@ class AuditResult:
     recommendations: List[str] = field(default_factory=list)
     claim: str = ""
     triggered_rules: List[str] = field(default_factory=list)
+    metadata: Dict[str, Any] = field(default_factory=dict)
 
     @property
     def ok(self) -> bool:
