+++ src/agt/version.py
@@ -1,0 +1,11 @@
+"""
+Canonical versioning for AGI-GAIA-TECHNE.
+"""
+
+AGT_REPO_VERSION = "8.8"
+FUNCTIONAL_CORE_VERSION = "4.3"
+CTK_VERSION = "4.2.2"
+CHK_VERSION = "0.3"
+PACKAGE_VERSION = "0.3.0"
+CANONICAL_ARCHITECTURE = "Unified Prism-Werk"
+LAST_UPDATED = "2026-05-24"
