+++ tests/test_canonical_axioms.py
@@ -1,0 +1,8 @@
+import pytest
+from agt.axioms import IS_WILLE, MACHINE_HAS_GEWISSEN, NO_GLOBAL_AUFHEBUNG, AGI_AS_TRANSCENDENTAL_HYPOTHESIS
+
+def test_axioms():
+    assert IS_WILLE is False
+    assert MACHINE_HAS_GEWISSEN is False
+    assert NO_GLOBAL_AUFHEBUNG is True
+    assert AGI_AS_TRANSCENDENTAL_HYPOTHESIS is True
